+++ modules/server/src/index.ts
@@ -9,3 +9,5 @@ export { FileContainer } from './internal/UploadFile';
 export { IServiceHandler } from './interfaces';
 export * from './ServerRequestWrapper';
 export * from './ServerResponseWrapper';
+export { createBeforeDecorator } from './utils';
+export { Status } from './enums';