+++ modules/server/src/interfaces/IInternalRoute.ts
@@ -9,4 +9,8 @@ export interface IInternalRoute extends IRoute {
      * The instance of the controller for this route.
      */
     controller: IController;
+    /**
+     * The target constructor so that we can use reflection inside the server.
+     */
+    target: any;
 }